requirements.txt
flask
flask_sqlalchemy
groq
python-dotenv
reportlab
requests

